Migrate `scraps tag` to `scraps tag list` subcommand
Convert tag from a flat command to a subcommand group following the
existing template pattern, preparing for future tag subcommands like
`tag backlinks`.

Closes #458
